requirements.txt
fastai==2.7.18
fastapi==0.116.0
numpy==1.24.4
numpy==1.21.5
opencv_python==4.10.0.84
Pillow==9.0.1
Pillow==11.3.0
ultralytics==8.3.162
